Adding in authentication groups in routes to keep non-authenticated users out of the routes they shouldn't be able to get to. Updating and adding ifs in views to make sure that even for authenticated users only admins can can see the controls (there's still a hole there because they can still use the routes, but I'm not too worried about it since it's supposed to be a personal blog at the moment).

+++ resources/views/pages/show.blade.php
@@ -5,17 +5,23 @@
 
     <div class="panel panel-default">
         <div class="panel-heading">
-                <h2>{{ $page->title  }} <small>{!! link_to_route('page.edit', "edit", ['page' => $page->slug]) !!}</small></h2>
+                <h2>{{ $page->title  }} 
+                    @if(Auth::check() && Auth::user()->administrator)
+                        <small>{!! link_to_route('page.edit', "edit", ['page' => $page->slug]) !!}</small>
+                    @endif
+                </h2>
         <div class="clearfix"></div>
         </div>
         <div class="panel-body">
             {{ $page->body }}
         </div>
-        <div class="panel-footer clearfix">
-            {!! Form::open(['method' => 'DELETE', 'route' => ['page.destroy', $page->id ]]) !!}
-                {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-            {!! Form::close() !!}
-        </div>
+        @if(Auth::check() && Auth::user()->administrator)
+            <div class="panel-footer clearfix">
+                {!! Form::open(['method' => 'DELETE', 'route' => ['page.destroy', $page->id ]]) !!}
+                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                {!! Form::close() !!}
+            </div>
+        @endif
     </div>
 
 @endsection
